read_rules() ignores empty lines and comments
Modified emod to pass the testcase in commit
2678fc0d9dd7ed8ddb1a0f8072f68dd349adec8a

read_rules() will ignore all lines that either start with a "#" or are
made up of whitespace (empty line). It also sorts the results before
returning.

The changes only apply with working with a file, not a directory since i
have not written a test for that yet.