Test save_rules.
Added a testcase for save_rules() when operating on a directory
structure, this test checks if the files are written correctly.

Currently this test uses read_rules() to verify the written files, and
read_rules() currently sorts the rules before returning them, which
means that the test will successed even if the rules are not sorted
before writting them.