Use new data folder, and test file_to_directory().
Created a Data folder that contains testfiles to be used in testcases,
Now each test methods that require these files will copy them from the
Data/ folder before actually using them, this allows us to reuse data
files with different names.

Added a 2 test methods for file_to_directory(). First check if it
creates anything and second check if the file it created is a directory.