Fix #11
argparse will now consider everything specified after the atom argument
as a use flag regardless of the syntax or the number, with this we
killed 2 birds with 1 stone. 1. this fixed issue #11 as argparse will no
longer try to interpret these flags as command-line switches and 2. this
made supplying use flags easier since it no longer requires the
`-e/--enable` switch.