Create Package class.
Created a Package() class that holds all of the package info and
describes it, i have modified the save_rules() and read_rules() to work
as methods for this class.

I did away with the seperate file_to_directory() and directory_to_file()
functions and instead now we have a a convert() method, it makes more
sense this way since there are only 2 package style so no need to
specify which one we need to convert to each time.

the convert method by itself only changes the Package.style variable
and and renames the package file, after which save_rules takes care of
the rest.

I have removed the author, email and date fields as they belong
somewhere else but kept the version and added a -v/--version
command-line argument that prints out the program name and its version.

Moved all testing files to the testing branch.